feat: move MCP server management to a new Plugins page
*This description was generated by Claude but is human-reviewed.*

Renames the "Hub" tab to "Templates" and adds a "Plugins" tab, the new
home for MCP server management ("plugin" is the user-facing name for an
MCP server). Configured servers show as cards, and the add-server form
now opens in a dialog. Everything moved as-is from Settings, so this
adds no new capability.

- The old ToolsMcpSettings component (~1000 lines) is split into
  focused components under src/components/plugins/: the list, the
  server card, the add dialog, and the shared key-value editor.
- Renderer-only: the MCP IPC handlers, OAuth flow, and database are
  untouched.
- The e2e tests get a Plugins page-object helper, and the MCP specs
  now navigate through the Plugins page instead of Settings.

This prepares for upcoming work to reduce the friction of adopting MCP
servers; doing the move now means that work lands on a proper page
instead of being built into Settings and moved later. Follow-ups:

1. dyad-cloud: expose a curated catalog of known MCP servers.
2. Show the catalog on the Plugins page: one-click add, with OAuth
   started automatically where needed.
3. Add an add_mcp_server tool so the AI can set up plugins too (always
   behind explicit user consent).

All MCP and template e2e specs updated and passing; typecheck, lint,
and unit tests clean.
