+++ e2e-tests/helpers/page-objects/components/index.ts
@@ -14,3 +14,4 @@ export { ModelPicker } from "./ModelPicker";
 export { Settings } from "./Settings";
 export { AppManagement } from "./AppManagement";
 export { PromptLibrary } from "./PromptLibrary";
+export { Plugins } from "./Plugins";
