+++ src/components/SettingsList.tsx
@@ -21,7 +21,6 @@ const SETTINGS_SECTIONS: SettingsSection[] = [
   { id: SECTION_IDS.telemetry, label: "Telemetry" },
   { id: SECTION_IDS.integrations, label: "Integrations" },
   { id: SECTION_IDS.agentPermissions, label: "Agent Permissions" },
-  { id: SECTION_IDS.toolsMcp, label: "Tools (MCP)" },
   { id: SECTION_IDS.advanced, label: "Advanced" },
   { id: SECTION_IDS.experiments, label: "Experiments" },
   { id: SECTION_IDS.dangerZone, label: "Danger Zone" },
