+++ src/components/app-sidebar-state.ts
@@ -7,7 +7,7 @@ export type AppSidebarHoverState =
 
 export type AppSidebarPanel = "Apps" | "Settings" | "Library";
 
-export type AppSidebarItemTitle = AppSidebarPanel | "Hub";
+export type AppSidebarItemTitle = AppSidebarPanel | "Templates" | "Plugins";
 
 export function getRouteSidebarPanel(pathname: string): AppSidebarPanel | null {
   if (
@@ -82,7 +82,10 @@ export function isSidebarItemActive({
   if (title === "Library") {
     return pathname.startsWith("/library");
   }
-  return pathname.startsWith("/hub");
+  if (title === "Templates") {
+    return pathname.startsWith("/templates");
+  }
+  return pathname.startsWith("/plugins");
 }
 
 export function shouldShowSelectedAppChatList({
