+++ src/components/app-sidebar.tsx
@@ -5,6 +5,7 @@ import {
   HelpCircle,
   Store,
   BookOpen,
+  Blocks,
 } from "lucide-react";
 import { Link, useNavigate, useRouterState } from "@tanstack/react-router";
 import { useSidebar } from "@/components/ui/sidebar"; // import useSidebar hook
@@ -59,9 +60,14 @@ const items = [
     icon: BookOpen,
   },
   {
-    title: "Hub",
-    to: "/hub",
+    title: "Templates",
+    to: "/templates",
     icon: Store,
+  },
+  {
+    title: "Plugins",
+    to: "/plugins",
+    icon: Blocks,
   },
 ] satisfies Array<{
   title: AppSidebarItemTitle;
@@ -296,8 +302,8 @@ function AppIcons({
       case "Library":
         return "start-hover:library";
       default:
-        // Items without a sub-list (e.g. Hub) dismiss any open preview so a
-        // stale list doesn't linger while hovering an unrelated icon.
+        // Items without a sub-list (Templates, Plugins) dismiss any open
+        // preview so a stale list doesn't linger while hovering an unrelated icon.
         return "clear-hover";
     }
   };
