+++ src/router.ts
@@ -5,7 +5,8 @@ import { chatRoute } from "./routes/chat";
 import { settingsRoute } from "./routes/settings";
 import { providerSettingsRoute } from "./routes/settings/providers/$provider";
 import { appDetailsRoute } from "./routes/app-details";
-import { hubRoute } from "./routes/hub";
+import { templatesRoute } from "./routes/templates";
+import { pluginsRoute } from "./routes/plugins";
 import { libraryRoute } from "./routes/library";
 import { appsRoute } from "./routes/apps";
 import { themesRoute } from "./routes/themes";
@@ -14,7 +15,8 @@ import { mediaRoute } from "./routes/media";
 
 const routeTree = rootRoute.addChildren([
   homeRoute,
-  hubRoute,
+  templatesRoute,
+  pluginsRoute,
   libraryRoute,
   appsRoute,
   themesRoute,
