+++ src/routes/templates.ts
@@ -1,0 +1,9 @@
+import { Route } from "@tanstack/react-router";
+import TemplatesPage from "../pages/templates";
+import { rootRoute } from "./root";
+
+export const templatesRoute = new Route({
+  getParentRoute: () => rootRoute,
+  path: "/templates",
+  component: TemplatesPage,
+});
